Unity Engine Barebone Library & Unity Test Project
Added Unity Project Folder with the Library built on osx and a example interface
Added gitignores for unity
Added ScreenCaptureInterface.h with C interface

limited to one window that can be received by unity as texture which is upside down.

the build-unity-library.sh requires the cmake .o files to generate a dylib. I have actually very little knowledge if that's a good way to generate the dylib, but it works on my osx computer.